requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.8.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-jazzmin==2.5.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.21.3
drf-yasg2==1.19.4
glob2==0.7
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
itypes==1.2.0
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
jasmine==3.99.0
jasmine-core==3.99.0
Jinja2==2.11.3
MarkupSafe==2.0.1
more-itertools==8.13.0
numpy==1.23.2
ordereddict==1.1
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
portend==3.1.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tempora==5.0.2
trio==0.21.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.11
whitenoise==6.2.0
wsproto==1.1.0
zc.lockfile==2.0
zipp==3.8.1
